Generate stamps for early classes
This fixes the build again. I can call (precompiled) generic
functions! Woo! I think that's the hardest part of this out of the
way.

The invalid FASL problem was that class-stamp-for-instances was
returning UNBOUND, which the FASL loader interpreted as the LTV
not being executed even though it actually was. What a pain.