Fix method functions
So they can actually be called. wowee

obviously this could stand to be more optimized.

It's necessary to actually use an interned symbol for the function
name since the functions are reference later in satiate, and
obviously the loader isn't going to construct the same uninterned
symbol twice. Could be finessed a bit I guess.